run filters itertools_imports -f itertools -f long -f map -f methodattrs -f ne -f next -f nonzero

+++ Lib/BaseHTTPServer.py
@@ -366,12 +366,12 @@ class BaseHTTPRequestHandler(socketserver.StreamRequestHandler):
         """
 
         try:
-            short, long = self.responses[code]
+            short, int = self.responses[code]
         except KeyError:
-            short, long = '???', '???'
+            short, int = '???', '???'
         if message is None:
             message = short
-        explain = long
+        explain = int
         self.log_error("code %d, message %s", code, message)
         # using _quote_html to prevent Cross Site Scripting attacks (see bug #1100201)
         content = (self.error_message_format %
